backend/requirements.txt
# Flask核心框架
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# 数据库
PyMySQL==1.1.0

# 请求处理
requests==2.31.0

# 文件处理
PyPDF2==3.0.1
python-docx==0.8.11

# 环境配置
python-dotenv==1.0.0

# 日志
loguru==0.7.2

# 讯飞SDK相关
websocket-client==1.6.4
websockets==12.0

# 其他依赖
Werkzeug==3.0.1
cryptography==41.0.7

# PDF处理
pdfplumber==0.10.3
pdf2image==1.16.3
Pillow==10.1.0

# WebSocket支持
flask-socketio==5.3.6
python-socketio[asyncio_client]==5.11.2

# 音频处理
pyaudio==0.2.11
pydub==0.25.1

# 图像处理和计算机视觉
opencv-python==4.8.1.78
pyautogui==0.9.54

# 数据处理
pandas==2.1.4
numpy==1.24.4

# 异步HTTP客户端
aiohttp==3.9.1

# 密码加密
bcrypt==4.1.2
